fix: move dto and types
