+++ backend/src/auth/dto/register.dto.ts
@@ -1,0 +1,13 @@
+import { IsEmail, IsNotEmpty, IsStrongPassword } from 'class-validator';
+
+export class RegisterDto {
+  @IsEmail()
+  email: string;
+
+  @IsNotEmpty()
+  name: string;
+
+  @IsNotEmpty()
+  @IsStrongPassword()
+  password: string;
+}
