+++ backend/src/auth/strategies/jwt/jwt-auth.guard.ts
