+++ backend/src/auth/strategies/jwt/jwt.strategy.ts
